Capitalize installer.txt symlink
I was deploying a Fork CMS 3.8 website with Capistrano, but when I visited the website, it kept redirecting to /install. I found out that Fork CMS verifies that `Installed.txt` exists, if not: he redirects to /install.

Forkcms_deploy creates a symlink `/src/Install/Cache/installed.txt` so when I capitalized this symlink, the website started working.