Use skip-lock-tables to avoid long table locks.
Some of our projects have rather big databases. Not locking these tables will make sure
our application doesn't have to wait for a long time during deployment.

A downside of this approach is the possibility of inconsistent data. During migrations, a
maintenance page is shown though, so this should not cause problems.